requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
langchain==0.0.350
langchain-openai==0.0.2
pydantic==2.5.2
python-multipart==0.0.6
markdown2==2.4.10
pygments==2.17.2 